feat: create Average Chart
